Update direction_screen.dart
Direction screen is changed and updated. We deleted alert box and add button for choosing role(Driver or Passenger) but cannot integrate on MyHomePage. We try to add choosing role button before the starting map. User should enter the map by choosing their role before the map opens and if they wants, they can change their role while the map is open. We are trying to integrate this.